Refactor Login and Register Forms into a Unified Component
- Created `LoginRegisterForm.tsx` to handle both login and registration forms
- Updated `LoginPage.tsx` to use `LoginRegisterForm` for dynamic form rendering
